back up

+++ code/competition/Codeforces Round #647/6.cpp
@@ -3,7 +3,7 @@ using namespace std;
 const int inf = 0x7FFFFFFF;
 typedef long long ll;
 #define int long long
-const ll mod = (0 ? 1000000007 : 998244353);
+const int mod = (0 ? 1000000007 : 998244353);
 const double eps = 1e-7;
 const ll llinf = 4223372036854775807;
 inline void out(int a)
